+++ R/globals_extra.R
@@ -1,4 +1,5 @@
 utils::globalVariables(c(
   "frac_optimal", "frac_LL", "response_rate", "breakpoint", "frac_risky",
-  "paradigm", "score", "reinforcement_rate", "delay", "v", "value", "name", "episode"
+  "paradigm", "score", "reinforcement_rate", "delay", "v", "value", "name", "episode",
+  "step", "responses", "elapsed"
 ))